#5 : UI started with landing page menu and Sandbox button. Some enhancements added as well

+++ lib/features/instruments/painters/rmi_painters.dart
@@ -136,7 +136,7 @@ class RMIRosePainter extends CustomPainter {
 }
 
 class RMIBugPainter extends CustomPainter {
-  final double targetHeading;
+  final int targetHeading;
 
   RMIBugPainter({required this.targetHeading});
